renamed file for user routes for consistency
